Generate instanceOf checks in Block Versioner for index field objects
When versioning blocks, it's possible for the generated checks to load
index variables from objects by duplicating code from within blocks that
are protected by inlining guards. Therefore the object might not be of
the expected type possibly resulting in loads that reach past the end
of the object. In rare cases this load can result in a SEGV when the
object is allocated at the very end of the heap.

Signed-off-by: Kevin Langman <langman@ca.ibm.com>